requirements.txt
torch
torchvision
pillow
ultralytics
scikit-learn
numpy
grad-cam
opencv-python
openai
python-dotenv
